refactor(cli): give organizations and environments their own resolvers
orgenv.go held resolution for both kinds behind a name that said nothing
about what it did. Split the two combined entry points so each kind's
lookup lives beside the client calls and error messages it already owns:

  ResolveNamedOrgAndEnv -> ResolveNamedOrg   (organizations.go)
                           ResolveNamedEnv   (environments.go)
  ResolveOrgAndEnvIDs   -> ResolveOrgOrPrompt (organizations.go)
                           ResolveEnvOrPrompt (environments.go)

The names now say what separates them: OrPrompt may fall through to the
interactive selector, Named never does. That also removes a pair that
differed from the exported ResolveOrgID/ResolveEnvID only by case while
doing something else entirely.

Splitting makes the ordering explicit rather than hidden in a function
body: environments are scoped to an organization, so the organization has
to resolve first and ResolveEnvOrPrompt takes the resolved id as an
argument. Callers say so in a comment, and orgIDFallback is now only on
the environment side, where it was ever meaningful.

Callers report the two steps separately, so a failure names which half
broke instead of "resolving organization and environment".

What is left of orgenv.go is the TTY gate, now selector.go. Tests follow
their subjects into organizations_test.go and environments_test.go, with
the Control Plane stub and the selector override in stubs_test.go.

One case is gone: "a bad org name does not fall through to the environment
lookup" tested the combined function's short-circuit, which is now the
caller's error check. Three arrive: an explicit id winning over a name at
each level, and a non-interactive terminal erroring per level.
